consent: migrate to test helpers [closes #1043]
